Fix unicode characters in strings between Lua -> Java and Java -> Lua
This is such a major hack in my opinion but it works.

The problems seems to be due to the fact that Java uses Modified UTF-8 which limits characters to three bytes (instead of four) while Lua uses normal UTF-8 so there's a small mismatch. This change will simply check if the string needs fixing and instead of passing the string directly through Java/JNI it will pass the raw bytes and construct a string from them